Consolidate parsing methods from ffi and non-ffi code (fixes #160) (#161)
* Move parsing methods from ffi to ir namespace.

* Replace existing parsing methods with those from ir namespace.

* Improve performance of parsing methods from ffi to match those in CLP.

* Fix build error

* C++ 101: Inline methods should be defined in the header.

* Replace LogTypeDictionaryEntry::VarDelim with ir::VariablePlaceholder; Deduplicate logtype constant escaping method; Replace all instances of variable delimiter with variable placeholder.

* Apply suggestions from code review


* Add suggestion from code review with formatting.

---------
