+++ components/core/src/utils/make_dictionaries_readable/README.md
@@ -4,6 +4,6 @@ For a dictionary, `make-dictionaries-readable` prints one entry per line.
 For log type dictionary entries, this requires making some characters printable:
 
 * Newlines are replaced with `\n`
-* Dictionary variable delimiters are replaced with `\d`
-* Non-dictionary integer variable delimiters are replaced with `\i`
-* Non-dictionary float variable delimiters are replaced with `\f`
+* Dictionary variable placeholders are replaced with `\d`
+* Non-dictionary integer variable placeholders are replaced with `\i`
+* Non-dictionary float variable placeholders are replaced with `\f`
