+++ components/core/tests/test-Utils.cpp
@@ -82,85 +82,3 @@ TEST_CASE("get_unambiguous_path", "[get_unambiguous_path]") {
     REQUIRE(get_unambiguous_path("abc///def/../ghi/./") == "abc/ghi");
     REQUIRE(get_unambiguous_path("../abc///def/../ghi/./") == "abc/ghi");
 }
-
-TEST_CASE("get_bounds_of_next_var", "[get_bounds_of_next_var]") {
-    string str;
-    size_t begin_pos;
-    size_t end_pos;
-
-    // Corner cases
-    // Empty string
-    str = "";
-    begin_pos = 0;
-    end_pos = 0;
-    REQUIRE(get_bounds_of_next_var(str, begin_pos, end_pos) == false);
-
-    // end_pos past the end of the string
-    str = "abc";
-    begin_pos = 0;
-    end_pos = string::npos;
-    REQUIRE(get_bounds_of_next_var(str, begin_pos, end_pos) == false);
-
-    // Non-variables
-    str = "/";
-    begin_pos = 0;
-    end_pos = 0;
-    REQUIRE(get_bounds_of_next_var(str, begin_pos, end_pos) == false);
-    REQUIRE(str.length() == begin_pos);
-
-    str = "xyz";
-    begin_pos = 0;
-    end_pos = 0;
-    REQUIRE(get_bounds_of_next_var(str, begin_pos, end_pos) == false);
-    REQUIRE(str.length() == begin_pos);
-
-    str = "=";
-    begin_pos = 0;
-    end_pos = 0;
-    REQUIRE(get_bounds_of_next_var(str, begin_pos, end_pos) == false);
-    REQUIRE(str.length() == begin_pos);
-
-    // Variables
-    str = "~=x!abc123;1.2%x:+394/-";
-    begin_pos = 0;
-    end_pos = 0;
-
-    REQUIRE(get_bounds_of_next_var(str, begin_pos, end_pos) == true);
-    REQUIRE("x" == str.substr(begin_pos, end_pos - begin_pos));
-
-    REQUIRE(get_bounds_of_next_var(str, begin_pos, end_pos) == true);
-    REQUIRE("abc123" == str.substr(begin_pos, end_pos - begin_pos));
-
-    REQUIRE(get_bounds_of_next_var(str, begin_pos, end_pos) == true);
-    REQUIRE("1.2" == str.substr(begin_pos, end_pos - begin_pos));
-
-    REQUIRE(get_bounds_of_next_var(str, begin_pos, end_pos) == true);
-    REQUIRE("+394/-" == str.substr(begin_pos, end_pos - begin_pos));
-
-    REQUIRE(get_bounds_of_next_var(str, begin_pos, end_pos) == false);
-
-    str = " ad ff 95 24 0d ff ";
-    begin_pos = 0;
-    end_pos = 0;
-
-    REQUIRE(get_bounds_of_next_var(str, begin_pos, end_pos) == true);
-    REQUIRE("ad" == str.substr(begin_pos, end_pos - begin_pos));
-
-    REQUIRE(get_bounds_of_next_var(str, begin_pos, end_pos) == true);
-    REQUIRE("ff" == str.substr(begin_pos, end_pos - begin_pos));
-
-    REQUIRE(get_bounds_of_next_var(str, begin_pos, end_pos) == true);
-    REQUIRE("95" == str.substr(begin_pos, end_pos - begin_pos));
-
-    REQUIRE(get_bounds_of_next_var(str, begin_pos, end_pos) == true);
-    REQUIRE("24" == str.substr(begin_pos, end_pos - begin_pos));
-
-    REQUIRE(get_bounds_of_next_var(str, begin_pos, end_pos) == true);
-    REQUIRE("0d" == str.substr(begin_pos, end_pos - begin_pos));
-
-    REQUIRE(get_bounds_of_next_var(str, begin_pos, end_pos) == true);
-    REQUIRE("ff" == str.substr(begin_pos, end_pos - begin_pos));
-
-    REQUIRE(get_bounds_of_next_var(str, begin_pos, end_pos) == false);
-    REQUIRE(str.length() == begin_pos);
-}
